move some files into a utils/ folder
