+++ src/camera.js
@@ -1,6 +1,6 @@
 /*global Camera */
 import Geo from './geo';
-import Utils from './utils';
+import Utils from './utils/utils';
 import ShaderProgram from './gl/shader_program';
 
 import glMatrix from 'gl-matrix';
